* Refactor events * Add BeforeRemoveBlockFromWorld
