+++ src/API/Event/Delegates/ActionDelegate.cs
@@ -1,6 +1,6 @@
 using Vintagestory.API.Common;
 
-namespace CarryOn.API.Event
+namespace CarryOn.API.Event.Delegates
 {
     public delegate void ActionDelegate(EnumEntityAction action, ref EnumHandling handled);
 }
