+++ src/API/Event/Delegates/BlockEntityDataDelegate.cs
@@ -1,7 +1,7 @@
 using Vintagestory.API.Common;
 using Vintagestory.API.Datastructures;
 
-namespace CarryOn.API.Event
+namespace CarryOn.API.Event.Delegates
 {
     public delegate void BlockEntityDataDelegate(BlockEntity blockEntity, ITreeAttribute blockEntityData, bool dropped = false);
 }
