+++ src/API/Event/Delegates/CheckPermissionToCarryDelegate.cs
@@ -1,7 +1,7 @@
 using Vintagestory.API.Common;
 using Vintagestory.API.MathTools;
 
-namespace CarryOn.API.Event
+namespace CarryOn.API.Event.Delegates
 {
     public delegate void CheckPermissionToCarryDelegate(EntityPlayer playerEntity, BlockPos pos, bool isReinforced, out bool? hasPermission);
 }
