+++ src/API/Event/Delegates/RemoveBlockDelegate.cs
@@ -1,0 +1,7 @@
+using CarryOn.API.Common.Models;
+using Vintagestory.API.MathTools;
+
+namespace CarryOn.API.Event.Delegates
+{
+    public delegate void RemoveBlockDelegate(CarriedBlock carriedBlock, BlockPos pos);
+}
